-Missile prefabs are derived base missile prefab. -New orange missile prefab has added. -Some bugs has solved. -Missiles now exploding when hitting ground. -

+++ Assets/Scripts/Data/ValueObject/MissileLevelData.cs
@@ -15,6 +15,6 @@ namespace Data.ValueObject
         public int MissileCount = 10;
         public List<GameObject> MissilePrefabList;
         public int MissileMaxVelocity = 10;
-        public int MissileCreateOffset = 2;
+        public float MissileCreateOffset = 2;
     }
 }
